# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  github_com_gsystes_backend_internal_communication_dto.AssignPermissionsRequest:
    properties:
      permission_ids:
        items:
          type: integer
        type: array
    required:
    - permission_ids
    type: object
...
  github_com_gsystes_backend_internal_communication_dto.CreateRoleRequest:
    properties:
      code:
        maxLength: 64
        type: string
      description:
        maxLength: 256
        type: string
      name:
        maxLength: 64
        type: string
    required:
    - code
    - name
    type: object
...
  github_com_gsystes_backend_internal_communication_dto.UpdateRoleRequest:
    properties:
      code:
        maxLength: 64
        type: string
      description:
        maxLength: 256
        type: string
      name:
        maxLength: 64
        type: string
      status:
        enum:
        - 1
        - 2
        type: integer
    required:
    - code
    - name
    type: object
...
  github_com_gsystes_backend_internal_infrastructure_utils.PageResponse:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  github_com_gsystes_backend_internal_infrastructure_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 中后台管理系统后端 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gsystes Backend API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: 分页查询角色列表
      parameters:
      - description: 页码（默认 1）
        in: query
        name: page
        type: integer
      - description: 每页条数（默认 10，最大 100）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.PageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
      security:
      - BearerAuth: []
      summary: 角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建一个新角色
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_gsystes_backend_internal_communication_dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
            - properties:
                data:
                  properties:
                    id:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - 角色管理
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定角色
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 根据 ID 获取角色详细信息和关联权限
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取角色详情
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 更新指定角色的信息
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_gsystes_backend_internal_communication_dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
      security:
      - BearerAuth: []
      summary: 更新角色
      tags:
      - 角色管理
  /roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: 获取指定角色已分配的权限列表
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取角色权限
      tags:
      - 角色-权限
    post:
      consumes:
      - application/json
      description: 为指定角色分配权限
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限 ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_gsystes_backend_internal_communication_dto.AssignPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
      security:
      - BearerAuth: []
      summary: 分配角色权限
      tags:
      - 角色-权限
  /roles/all:
    get:
      consumes:
      - application/json
      description: 返回所有角色列表，只包含 ID、名称和编码
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gsystes_backend_internal_infrastructure_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取所有角色（精简）
      tags:
      - 角色管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"